fix: update min avalable date for the datepicer, update styles; refactor: update file name for the city selectm update name for functions
