+++ src/components/CitySelect/CitySelectStyles.scss
@@ -16,6 +16,7 @@
     line-height: 24px;
     cursor: pointer;
     position: relative;
+    margin-right: 0.5em;
 
     &:hover {
         border: 2px solid $accent-color;
@@ -56,5 +57,11 @@
 
     &:hover {
         background: $secondary-06;
+    }
+}
+
+@media screen and (max-device-width: 600px) {
+    .select-city {
+        margin-bottom: 1em;
     }
 }
