+++ src/components/Datepicker/Datepicker.tsx
@@ -1,13 +1,14 @@
 import React, { useState } from 'react';
 
 interface IProps {
-    getDate: (date: Date | string) => void;
+    changeDate: (date: Date | string) => void;
 }
 
-const Datepicker: React.FC<IProps> = ({ getDate }) => {
+const Datepicker: React.FC<IProps> = ({ changeDate }) => {
     const [toggleCalendarIcon, setToggleCalendarIcon] = useState<boolean>(true);
     const [pastDate, setPastDate] = useState<string>('Select date');
     const today: Date = new Date();
+    const countDaysRange = 4;
     today.setDate(today.getDate() - 1);
 
     const getMonth = (todayMonth: Date): string => {
@@ -15,21 +16,23 @@ const Datepicker: React.FC<IProps> = ({ getDate }) => {
     };
 
     const maxDatepicerDate = `${today.getFullYear()}-${getMonth(today)}-${today.getDate()}`;
-    today.setDate(today.getDate() - 5);
+    today.setDate(today.getDate() - countDaysRange);
     const minDatepicerDate = `${today.getFullYear()}-${getMonth(today)}-${today.getDate()}`;
 
     const handleDateChange = (event: React.ChangeEvent<HTMLInputElement>): void => {
-        getDate(event.target.value);
+        changeDate(event.target.value);
         const selectedDate = new Date(event.target.value);
         setPastDate(`${getMonth(selectedDate)}/${selectedDate.getDate()}/${selectedDate.getFullYear()}`);
+        handleInputDateTypeChange(event);
     };
 
-    const handleInputTextTypeChange = (event: React.MouseEvent<HTMLInputElement> | React.FocusEvent<HTMLInputElement>): void => {
+    const handleInputTextTypeChange = (event: React.FocusEvent<HTMLInputElement>): void => {
         event.currentTarget.type = 'date';
+        event.currentTarget.placeholder = 'mm/dd/yyy';
         setToggleCalendarIcon(false);
     };
 
-    const handleInputDateTypeChange = (event: React.MouseEvent<HTMLInputElement> | React.FocusEvent<HTMLInputElement>): void => {
+    const handleInputDateTypeChange = (event: React.FocusEvent<HTMLInputElement> | React.ChangeEvent<HTMLInputElement>): void => {
         event.currentTarget.type = 'text';
         event.currentTarget.placeholder = 'Select date';
         setToggleCalendarIcon(true);
@@ -42,7 +45,7 @@ const Datepicker: React.FC<IProps> = ({ getDate }) => {
 
     return (
         <div className='select-date'>
-            <input type='text' max={maxDatepicerDate} min={minDatepicerDate} name='select-date' onMouseOver={handleInputTextTypeChange} onBlur={handleInputDateTypeChange} onMouseOut={handleInputDateTypeChange} value={pastDate} onChange={handleDateChange}></input>
+            <input type='text' max={maxDatepicerDate} min={minDatepicerDate} name='select-date' onFocus={handleInputTextTypeChange} onBlur={handleInputDateTypeChange} value={pastDate} onChange={handleDateChange}></input>
             {toggleCalendarIcon ? <div className="select-date__img-datepicker" onClick={changeInputTextTypeByClick}></div> : null}
         </div>
     );
