+++ src/components/NearestWeather/NearestWeatherStyles.scss
@@ -9,6 +9,7 @@
     background-color: $weather-container-background;
     padding-bottom: 50px;
     box-shadow: 0px 4px 4px rgba(4, 5, 73, 0.25), 14px 14px 20px rgba(5, 6, 114, 0.2);
+    margin-right: 1em;
 }
 
 .forecast-container__details {
@@ -27,9 +28,10 @@
    }
 }
 
-@media screen and (max-device-width: 600px) {
-   .forecast-container__card {
-        padding: 0px 25px;
+@media screen and (max-device-width: 1280px) {
+    .forecast-container {
+        margin-right: 0em;
+        margin-bottom: 1em;
     }
 }
 
@@ -40,6 +42,10 @@
 }
 
 @media screen and (max-device-width: 600px) {
+   .forecast-container__card {
+        padding: 0px 25px;
+    }
+
     .forecast-container {
         width: 300px;
     }
