+++ src/components/WeatherCard/WeatherCardStyles.scss
@@ -17,7 +17,7 @@
     font-weight: bold;
     font-size: 16px;
     line-height: 24px;
-    margin-bottom: 15;
+    margin-bottom: 15px;
 }
 
 .card-container__temp {
